Added CellHub static class to encapsulate SignalR functionality
