+++ MobileTag/MobileTag.csproj
@@ -136,6 +136,7 @@
     <Compile Include="SharedCode\ColorCode.cs" />
     <Compile Include="SharedCode\Database.cs" />
     <Compile Include="SharedCode\MapOverlay.cs" />
+    <Compile Include="SharedCode\CellHub.cs" />
     <Compile Include="SplashScreen.cs" />
   </ItemGroup>
   <ItemGroup>
